requirements.txt
streamlit == 0.87
pandas == 1.2.4
st_aggridstreamlit-aggrid